STCOM-610: Clean Up
