+++ lib/FilterGroups/tests/FilterGroups-test.js
@@ -8,19 +8,21 @@ import { expect } from 'chai';
 
 import { mount } from '../../../tests/helpers';
 import FilterGroupInteractor from './interactor';
-import FilterGroupsSecondComponent from '../stories/filterGroupsUsageTest2';
-import FilterGroupsThirdComponent from '../stories/filterGroupsUsageTest3';
-import FilterGroupsFourthComponent from '../stories/filterGroupsUsageTest4';
-import FilterGroupsFifthComponent from '../stories/filterGroupsUsageTest5';
-import FilterGroupsSixthComponent from '../stories/filterGroupsUsageTest6';
+import {
+  FilterGroupsFirstComponent,
+  FilterGroupsSecondComponent,
+  FilterGroupsThirdComponent,
+  FilterGroupsFourthComponent,
+  FilterGroupsFifthComponent,
+} from './helpers/filterGroupsUsageTest';
 
-describe.only('Filter Group', () => {
+describe('Filter Group', () => {
   const filterGroupInteractor = new FilterGroupInteractor();
 
-  describe('render Filter Group Second component', () => {
+  describe('render Filter Group 1 component', () => {
     beforeEach(async () => {
       await mount(
-        <FilterGroupsSecondComponent />
+        <FilterGroupsFirstComponent/>
       );
     });
 
@@ -74,10 +76,10 @@ describe.only('Filter Group', () => {
       });
     });
   });
-  describe('render Filter Group Third component', () => {
+  describe('render Filter Group 2 component', () => {
     beforeEach(async () => {
       await mount(
-        <FilterGroupsThirdComponent />
+        <FilterGroupsSecondComponent />
       );
     });
 
@@ -130,10 +132,10 @@ describe.only('Filter Group', () => {
     });
   });
 
-  describe('render Filter Group Fourth component', () => {
+  describe('render Filter Group 3 component', () => {
     beforeEach(async () => {
       await mount(
-        <FilterGroupsFourthComponent />
+        <FilterGroupsThirdComponent />
       );
     });
 
@@ -168,10 +170,10 @@ describe.only('Filter Group', () => {
     });
   });
 
-  describe('render Filter Group Fifth component', () => {
+  describe('render Filter Group 4 component', () => {
     beforeEach(async () => {
       await mount(
-        <FilterGroupsFifthComponent />
+        <FilterGroupsFourthComponent />
       );
     });
 
@@ -217,10 +219,10 @@ describe.only('Filter Group', () => {
       });
     });
   });
-  describe('render Filter Group Sixth component', () => {
+  describe('render Filter Group 5 component', () => {
     beforeEach(async () => {
       await mount(
-        <FilterGroupsSixthComponent />
+        <FilterGroupsFifthComponent />
       );
     });
