+++ lib/FilterGroups/tests/interactor.js
@@ -14,7 +14,6 @@ export default interactor(class FilterGroupInteractor {
   hasControlGroup = isPresent('[data-test-filter-control-group]');
   checkbox = new CheckboxInteractor('[data-test-filter-control-group]');
   hasFilterCheckbox = isPresent('[data-test-filter-checkbox]');
-  clickFilterCheckbox = clickable('[data-test-filter-checkbox]');
   checkboxes = collection('[data-test-checkbox]', CheckboxInteractor);
   isFilterGroupVisible = isVisible('[data-test-accordion-wrapper]');
   hasClearButton = isPresent('[data-test-clear-button]');
